pyproject.toml
[project]
name = "scriptworker"
version = "62.2.1"
description = "Taskcluster scriptworker"
authors = [
    { name = "Mozilla Release Engineering", email = "releng@mozilla.com" },
]
readme = "README.rst"
requires-python = ">=3.11"
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "aiohttp>=3",
    "arrow>=1.0",
    "cryptography>=2.6.1",
    "dictdiffer",
    "github3.py",
    "immutabledict>=1.3.0",
    "jsonschema[format-nongpl]",
    "json-e>=2.5.0",
    "PyYAML",
    "taskcluster>=40",
    "taskcluster-taskgraph",
]

[dependency-groups]
docs = [
    "auxlib",
    "commonmark",
    "myst-parser",
    "recommonmark",
    # https://github.com/readthedocs/readthedocs.org/issues/10279
    "Sphinx<7",
    "sphinx_rtd_theme",
]
dev = [
    "tox",
    "tox-uv",
    "aiohttp>=3",
    "asyncio_extras",
    "black",
    "coveralls",
    "flake8",
    "flake8_docstrings",
    "isort>=5",
    "mock",
    "pydocstyle==3.0.0",
    "pytest",
    "pytest-asyncio<1.0",
    "pytest-mock",
    "pytest-random-order",
    "mypy",
    # For mypy stub discovery
    "aiohttp",
    "async_timeout",
    "types-PyYAML",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "taskcluster",
    ".readthedocs.yml",
    ".taskcluster.yml",
    "CODE_OF_CONDUCT.md",
    "docs",
    "mypi.ini",
    "tests",
    ".github",
]

[tool.distutils.bdist_wheel]
universal = true

[project.scripts]
scriptworker = "scriptworker.worker:main"
verify_cot = "scriptworker.cot.verify:verify_cot_cmdln"
create_test_workdir = "scriptworker.cot.verify:create_test_workdir"
verify_ed25519_signature = "scriptworker.ed25519:verify_ed25519_signature_cmdln"

[tool.black]
line-length = 160
target-version = ["py311", "py312", "py313"]
include = '\.(wsgi|pyi?)$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.cache
  | \.cache_py3
  | _build
  | buck-out
  | build
  | dist
  | ui
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 160

[tool.pytest.ini_options]
asyncio_mode = "auto"
